move secret board into own file
