+++ lib/heist_definitions.dart
@@ -48,16 +48,22 @@ const Map<int, Map<int, HeistDefinition>> heistDefinitions = {
     2: const HeistDefinition(numPlayers: 3, price: 20, maximumBid: 6),
     3: const HeistDefinition(numPlayers: 4, price: 24, maximumBid: 7),
     4: const HeistDefinition(numPlayers: 3, price: 20, maximumBid: 6),
-    5: const HeistDefinition(numPlayers: 4, price: 24, maximumBid: 7),
-  },
-  7: {
-    1: const HeistDefinition(numPlayers: 2, price: 14, maximumBid: 5),
-    2: const HeistDefinition(numPlayers: 3, price: 22, maximumBid: 6),
-    3: const HeistDefinition(numPlayers: 4, price: 26, maximumBid: 7),
-    4: const HeistDefinition(numPlayers: 3, price: 22, maximumBid: 6),
     5: const HeistDefinition(numPlayers: 4, price: 26, maximumBid: 7),
   },
-  8: {},
+  7: {
+    1: const HeistDefinition(numPlayers: 2, price: 12, maximumBid: 5),
+    2: const HeistDefinition(numPlayers: 3, price: 20, maximumBid: 6),
+    3: const HeistDefinition(numPlayers: 4, price: 24, maximumBid: 7),
+    4: const HeistDefinition(numPlayers: 3, price: 20, maximumBid: 6),
+    5: const HeistDefinition(numPlayers: 4, price: 26, maximumBid: 7),
+  },
+  8: {
+    1: const HeistDefinition(numPlayers: 3, price: 14, maximumBid: 6),
+    2: const HeistDefinition(numPlayers: 4, price: 24, maximumBid: 7),
+    3: const HeistDefinition(numPlayers: 4, price: 24, maximumBid: 7),
+    4: const HeistDefinition(numPlayers: 5, price: 26, maximumBid: 8),
+    5: const HeistDefinition(numPlayers: 5, price: 28, maximumBid: 8),
+  },
   9: {},
   10: {},
 };
